requirements.txt
django
djangorestframework
pytest
requests
